+++ huffman.h
@@ -13,17 +13,10 @@
 #include <stdlib.h>
 #include <string.h>
 
-/* Maximum size of the histogram table. */
-#define MAX_HIST_TAB_LEN (0x1U << 8)
 
 /* Maximum size of a encoded (or decoded) byte. */
 #define MAX_INT_BUF_BITS (0x8U)
 
-/*
- * Intermediate nodes in the tree are populated with
- * this byte to distinguish them from leaf nodes.
- */
-#define PSEUDO_TREE_BYTE (MAX_HIST_TAB_LEN - 0x1U)
 
 /*
  * Denotes the end of file (EOF) for our encoding or
@@ -37,35 +30,6 @@ typedef struct meta {
     uint64_t nr_bytes; /* Number of encoded bytes (excluding headers). */
     uint64_t nr_bits; /* Number of encoded bits (excluding headers). */
 } meta_t;
-
-/* Maps a character to its frequency in the data. */
-typedef struct map {
-    uint8_t ch;
-    uint32_t freq;
-} map_t;
-
-/* Represents a node in the priority queue (or tree). */
-typedef struct node {
-    struct node *next;  /* Next link in the linked list (queue). */
-    struct node *right; /* Tree node to the right. */
-    struct node *left;  /* Tree node to the left. */
-    struct map data;    /* A mapping of byte to its frequency. */
-} node_t;
-
-/* Routines for trees. */
-uint32_t tree_height(struct node *root);
-uint8_t tree_leaf(struct node *node);
-void traverse_tree(uint8_t, struct node *, int8_t, uint8_t *, int8_t *);
-void make_tree(struct node **);
-void nuke_tree(struct node **);
-
-/* Routines for queues. */
-uint32_t queue_size(struct node *);
-void enqueue(struct node **head, struct node *);
-struct node *dequeue(struct node **);
-void init_queue(struct node **);
-struct node *make_queue(struct map *, uint32_t);
-void nuke_queue(struct node **);
 
 /* Routines for maps. */
 uint32_t table_size(uint32_t *);
